Moved osgViewer::View _viewer to class OMVisViewer.
The view holds a osg scene and is part of the MainWindow object, thus it is created once and handled by the class OMVisViewer. If a model is loaded, we use this view to show the scene.
Furthermore, the osgViewer::View _viewer in class OMVisualViewer can be removed.

Renamed files according to class OMVisViewer.