Fixed bug: Successively loading several models into OMVis caused seg faults.
The bug was caused beaucse the load mechanism was not implemented with care. The method GuiController::loadModel now first checks, if the new model can be succesfully loaded and than the _omVisualizer attribute points to the new model. Otherwise the old model is kept.

Several methods now have a return value that indicates an error status.